Implement Github.pull_requests as a stream
the intent is that we only need to process the stream as far as the
oldest Jira/PR it needs to find.  That is, for each Jira ticket key
it can scan the PRs until it finds a match.  Most often, it won't
need to scan all PRs. Instead, we'll build a map, and when the map
doesn't have a match, it will continue scanning the stream and
populating the map.