Move state tracking for pull request paging into an Agent
Created GithubPRMap as an agent maintaining state of the cursor
for the paginated list of pull requests.  The idea is to only fetch
more PRs from github if needed, and keep a map (cache) of the ones
already fetched.

Before, i'd thought the Stream i produced
would somehow track the state of where it
was.  But, that was misguided.  Instead,
i needed to track state based on the cursor
and requery the graphql API with the cursor
when more data was needed.