Implement annotate_pull_request
Still having some problems with the graphql query.  On some repos,
like vendini-api, i can retrieve the PR's headRef but on others (ct-api)
graphql returns an error (that they concede is a bug).

Ideally, this should check the git SHA of the merged PR with the SHA
of the approval.

Also, there's an issue still with the case of a single JIRA ticket having
multiple PRs.  Currently, only one gets considered.  Should change the
map from jira key to pr to map to a list of PRs.  But, then what to do?